refactor: move search songs query to a isolated function
